separation of concerns, moved todo list into its own module, added AddTodoForm
